Move cancelable purchases selector
